requirements.txt
tensorflow-gpu
keras
profilehooks
numpy
pyperclip
python-chess
h5py